Crate folding_schemes

Modules

arith

commitment

constants

folding

frontend

transcript

utils

Enums

Error

Traits

Curve

Curve trait is a wrapper around CurveGroup that also includes the necessary bounds for the curve to be used conveniently in folding schemes.

Decider

Field

Field trait is a wrapper around PrimeField that also includes the necessary bounds for the field to be used conveniently in folding schemes.

FoldingScheme

FoldingScheme defines trait that is implemented by the diverse folding schemes. It is defined over a cycle of curves (C1, C2), where:

MultiFolding

Trait with auxiliary methods for multi-folding schemes (ie. HyperNova, ProtoGalaxy, etc), allowing to create new instances for the multifold.