folding_schemes::arith::ccs

Struct CCS

pub struct CCS<F: PrimeField> {
    pub t: usize,
    pub s: usize,
    pub M: Vec<SparseMatrix<F>>,
    pub S: Vec<Vec<usize>>,
    pub c: Vec<F>,
    /* private fields */
}

CCS represents the Customizable Constraint Systems structure defined in the CCS paper

Fields
t: usize

t = |M|, number of matrices

s: usize

s = log(m), dimension of x

M: Vec<SparseMatrix<F>>

vector of matrices

S: Vec<Vec<usize>>

vector of multisets

c: Vec<F>

vector of coefficients

Implementations

impl<F: PrimeField> CCS<F>

pub fn eval_at_z(&self, z: &[F]) -> Result<Vec<F>, Error>

Evaluates the CCS relation at a given vector of assignments z

impl<F: PrimeField> CCS<F>

pub fn to_cccs<R: Rng, C, CS: CommitmentScheme<C, H>, const H: bool>( &self, rng: &mut R, cs_params: &CS::ProverParams, z: &[F], ) -> Result<(CCCS<C>, Witness<F>), Error>
where C: Curve<ScalarField = F>,

pub fn compute_q(&self, z: &[F]) -> Result<VirtualPolynomial<F>, Error>

Computes q(x) = \sum^q c_i * \prod_{j \in S_i} ( \sum_{y \in {0,1}^s’} M_j(x, y) * z(y) ) polynomial over x

pub fn compute_Q( &self, z: &[F], beta: &[F], ) -> Result<VirtualPolynomial<F>, Error>

Computes Q(x) = eq(beta, x) * q(x) = eq(beta, x) * \sum^q c_i * \prod_{j \in S_i} ( \sum_{y \in {0,1}^s’} M_j(x, y) * z(y) ) polynomial over x

impl<F: PrimeField> CCS<F>

pub fn to_lcccs<R: Rng, C, CS: CommitmentScheme<C, H>, const H: bool>( &self, rng: &mut R, cs_params: &CS::ProverParams, z: &[F], ) -> Result<(LCCCS<C>, Witness<F>), Error>
where C: Curve<ScalarField = F>,