folding_schemes::arith

Trait Arith

pub trait Arith: Clone {
    // Required methods
    fn degree(&self) -> usize;
    fn n_constraints(&self) -> usize;
    fn n_variables(&self) -> usize;
    fn n_public_inputs(&self) -> usize;
    fn n_witnesses(&self) -> usize;
    fn split_z<F: PrimeField>(&self, z: &[F]) -> (Vec<F>, Vec<F>);
}

Arith is a trait about constraint systems (R1CS, CCS, etc.), where we define methods for getting information about the constraint system.

Required Methods

fn degree(&self) -> usize

Returns the degree of the constraint system

fn n_constraints(&self) -> usize

Returns the number of constraints in the constraint system

fn n_variables(&self) -> usize

Returns the number of variables in the constraint system

fn n_public_inputs(&self) -> usize

Returns the number of public inputs / public IO / instances / statements in the constraint system

fn n_witnesses(&self) -> usize

Returns the number of witnesses / secret inputs in the constraint system

fn split_z<F: PrimeField>(&self, z: &[F]) -> (Vec<F>, Vec<F>)

Returns a tuple containing (w, x) (witness and public inputs respectively)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.