folding_schemes::arith

Trait ArithRelation

pub trait ArithRelation<W, U>: Arith {
    type Evaluation;

    // Required methods
    fn eval_relation(&self, w: &W, u: &U) -> Result<Self::Evaluation, Error>;
    fn check_evaluation(w: &W, u: &U, v: Self::Evaluation) -> Result<(), Error>;

    // Provided methods
    fn dummy_witness_instance<'a>(&'a self) -> (W, U)
       where W: Dummy<&'a Self>,
             U: Dummy<&'a Self> { ... }
    fn check_relation(&self, w: &W, u: &U) -> Result<(), Error> { ... }
}

ArithRelation treats a constraint system as a relation between a witness of type W and a statement / public input / public IO / instance of type U, and in this trait, we define the necessary operations on the relation.

Note that the same constraint system may support different types of W and U, and the satisfiability check may vary.

For example, both plain R1CS and relaxed R1CS are represented by 3 matrices, but the types of W and U are different:

The plain R1CS has W and U as vectors of field elements.

W = w and U = x satisfy R1CS if Az ∘ Bz = Cz, where z = [1, x, w].

In Nova, Relaxed R1CS has W as crate::folding::nova::Witness, and U as crate::folding::nova::CommittedInstance.

W = (w, e, ...) and U = (u, x, ...) satisfy Relaxed R1CS if Az ∘ Bz = uCz + e, where z = [u, x, w]. (commitments in U are not checked here)

Also, W and U have non-native field elements as their components when used as CycleFold witness and instance.

In ProtoGalaxy, Relaxed R1CS has W as crate::folding::protogalaxy::Witness, and U as crate::folding::protogalaxy::CommittedInstance.

W = (w, ...) and U = (x, e, β, ...) satisfy Relaxed R1CS if e = Σ pow_i(β) v_i, where v = Az ∘ Bz - Cz, z = [1, x, w]. (commitments in U are not checked here)

This is also the case of CCS, where W and U may be vectors of field elements, crate::folding::hypernova::Witness and crate::folding::hypernova::lcccs::LCCCS, or crate::folding::hypernova::Witness and crate::folding::hypernova::cccs::CCCS.

Required Associated Types

type Evaluation

Required Methods

fn eval_relation(&self, w: &W, u: &U) -> Result<Self::Evaluation, Error>

Evaluates the constraint system self at witness w and instance u. Returns the evaluation result.

The evaluation result is usually a vector of field elements. For instance:

Evaluating the plain R1CS at W = w and U = x returns Az ∘ Bz - Cz, where z = [1, x, w].

Evaluating the relaxed R1CS in Nova at W = (w, e, ...) and U = (u, x, ...) returns Az ∘ Bz - uCz, where z = [u, x, w].

Evaluating the relaxed R1CS in ProtoGalaxy at W = (w, ...) and U = (x, e, β, ...) returns Az ∘ Bz - Cz, where z = [1, x, w].

However, we use Self::Evaluation to represent the evaluation result for future extensibility.

fn check_evaluation(w: &W, u: &U, v: Self::Evaluation) -> Result<(), Error>

Checks if the evaluation result is valid. The witness w and instance u are also parameters, because the validity check may need information contained in w and/or u.

For instance:

The evaluation v of plain R1CS at satisfying W and U should be an all-zero vector.

The evaluation v of relaxed R1CS in Nova at satisfying W and U should be equal to the error term e in the witness.

The evaluation v of relaxed R1CS in ProtoGalaxy at satisfying W and U should satisfy e = Σ pow_i(β) v_i, where e is the error term in the committed instance.

Provided Methods

fn dummy_witness_instance<'a>(&'a self) -> (W, U)
where W: Dummy<&'a Self>, U: Dummy<&'a Self>,

Returns a dummy witness and instance

fn check_relation(&self, w: &W, u: &U) -> Result<(), Error>

Checks if witness w and instance u satisfy the constraint system self by first computing the evaluation result and then checking the validity of the evaluation result.

Used only for testing.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.