folding_schemes::arith

Trait ArithRelationGadget

pub trait ArithRelationGadget<WVar, UVar> {
    type Evaluation;

    // Required methods
    fn eval_relation(
        &self,
        w: &WVar,
        u: &UVar,
    ) -> Result<Self::Evaluation, SynthesisError>;
    fn enforce_evaluation(
        w: &WVar,
        u: &UVar,
        e: Self::Evaluation,
    ) -> Result<(), SynthesisError>;

    // Provided method
    fn enforce_relation(&self, w: &WVar, u: &UVar) -> Result<(), SynthesisError> { ... }
}

ArithRelationGadget defines the in-circuit counterparts of operations specified in ArithRelation on constraint systems.

Required Associated Types

type Evaluation

Required Methods

fn eval_relation( &self, w: &WVar, u: &UVar, ) -> Result<Self::Evaluation, SynthesisError>

Evaluates the constraint system self at witness w and instance u. Returns the evaluation result.

fn enforce_evaluation( w: &WVar, u: &UVar, e: Self::Evaluation, ) -> Result<(), SynthesisError>

Generates constraints for enforcing that the evaluation result is valid. The witness w and instance u are also parameters, because the validity check may need information contained in w and/or u.

Provided Methods

fn enforce_relation(&self, w: &WVar, u: &UVar) -> Result<(), SynthesisError>

Generates constraints for enforcing that witness w and instance u satisfy the constraint system self by first computing the evaluation result and then checking the validity of the evaluation result.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.