folding_schemes::arith

Trait ArithSampler

pub trait ArithSampler<C: Curve, W, U>: ArithRelation<W, U> {
    // Required method
    fn sample_witness_instance<CS: CommitmentScheme<C, true>>(
        &self,
        params: &CS::ProverParams,
        rng: impl RngCore,
    ) -> Result<(W, U), Error>;
}

ArithSampler allows sampling random pairs of witness and instance that satisfy the constraint system self.

This is useful for constructing a zero-knowledge layer for a folding-based IVC. An example of such a layer can be found in Appendix D of the HyperNova paper.

Note that we use a separate trait for sampling, because this operation may not be supported by all witness-instance pairs. For instance, it is difficult (if not impossible) to do this for w and x in a plain R1CS.

Required Methods

fn sample_witness_instance<CS: CommitmentScheme<C, true>>( &self, params: &CS::ProverParams, rng: impl RngCore, ) -> Result<(W, U), Error>

Samples a random witness and instance that satisfy the constraint system.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.