folding_schemes::arith

Trait ArithSerializer

pub trait ArithSerializer {
    // Required method
    fn params_to_le_bytes(&self) -> Vec<u8> ;
}

ArithSerializer is for serializing constraint systems.

Currently we only support converting parameters to bytes, but in the future we may consider implementing methods for serializing the actual data (e.g., R1CS matrices).

Required Methods

fn params_to_le_bytes(&self) -> Vec<u8>

Returns the bytes that represent the parameters, that is, the matrices sizes, the amount of public inputs, etc, without the matrices/polynomials values.