folding_schemes::commitment

Trait CommitmentScheme

pub trait CommitmentScheme<C: Curve, const H: bool = false>: Clone + Debug {
    type ProverParams: Clone + Debug + CanonicalSerialize + CanonicalDeserialize;
    type VerifierParams: Clone + Debug + CanonicalSerialize + CanonicalDeserialize;
    type Proof: Clone + Debug + CanonicalSerialize + CanonicalDeserialize;
    type ProverChallenge: Clone + Debug;
    type Challenge: Clone + Debug;

    // Required methods
    fn is_hiding() -> bool;
    fn setup(
        rng: impl RngCore,
        len: usize,
    ) -> Result<(Self::ProverParams, Self::VerifierParams), Error>;
    fn commit(
        params: &Self::ProverParams,
        v: &[C::ScalarField],
        blind: &C::ScalarField,
    ) -> Result<C, Error>;
    fn prove(
        params: &Self::ProverParams,
        transcript: &mut impl Transcript<C::ScalarField>,
        cm: &C,
        v: &[C::ScalarField],
        blind: &C::ScalarField,
        rng: Option<&mut dyn RngCore>,
    ) -> Result<Self::Proof, Error>;
    fn prove_with_challenge(
        params: &Self::ProverParams,
        challenge: Self::ProverChallenge,
        v: &[C::ScalarField],
        blind: &C::ScalarField,
        rng: Option<&mut dyn RngCore>,
    ) -> Result<Self::Proof, Error>;
    fn verify(
        params: &Self::VerifierParams,
        transcript: &mut impl Transcript<C::ScalarField>,
        cm: &C,
        proof: &Self::Proof,
    ) -> Result<(), Error>;
    fn verify_with_challenge(
        params: &Self::VerifierParams,
        challenge: Self::Challenge,
        cm: &C,
        proof: &Self::Proof,
    ) -> Result<(), Error>;
}

CommitmentScheme defines the vector commitment scheme trait. Where H indicates if to use the commitment in hiding mode or not.

Required Associated Types

type ProverParams: Clone + Debug + CanonicalSerialize + CanonicalDeserialize

type VerifierParams: Clone + Debug + CanonicalSerialize + CanonicalDeserialize

type Proof: Clone + Debug + CanonicalSerialize + CanonicalDeserialize

type ProverChallenge: Clone + Debug

type Challenge: Clone + Debug

Required Methods

fn is_hiding() -> bool

fn setup( rng: impl RngCore, len: usize, ) -> Result<(Self::ProverParams, Self::VerifierParams), Error>

fn commit( params: &Self::ProverParams, v: &[C::ScalarField], blind: &C::ScalarField, ) -> Result<C, Error>

fn prove( params: &Self::ProverParams, transcript: &mut impl Transcript<C::ScalarField>, cm: &C, v: &[C::ScalarField], blind: &C::ScalarField, rng: Option<&mut dyn RngCore>, ) -> Result<Self::Proof, Error>

fn prove_with_challenge( params: &Self::ProverParams, challenge: Self::ProverChallenge, v: &[C::ScalarField], blind: &C::ScalarField, rng: Option<&mut dyn RngCore>, ) -> Result<Self::Proof, Error>

same as prove but instead of providing a Transcript to use, providing the already computed challenge

fn verify( params: &Self::VerifierParams, transcript: &mut impl Transcript<C::ScalarField>, cm: &C, proof: &Self::Proof, ) -> Result<(), Error>

fn verify_with_challenge( params: &Self::VerifierParams, challenge: Self::Challenge, cm: &C, proof: &Self::Proof, ) -> Result<(), Error>

same as verify but instead of providing a Transcript to use, providing the already computed challenge

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.