folding_schemes::folding

Module circuits

Modules

cyclefold

decider

nonnative

sum_check

utils

Type Aliases

CF1

CF1 uses the ScalarField of the given C. CF1 represents the ConstraintField used for the main folding circuit which is over E1::Fr, where E1 is the main curve where we do the folding. In CF1, the points of C can not be natively represented.

CF2

CF2 uses the BaseField of the given C. CF2 represents the ConstraintField used for the CycleFold circuit which is over E2::Fr=E1::Fq, where E2 is the auxiliary curve (from CycleFold approach) where we check the folding of the commitments (elliptic curve points). In CF2, the points of C can be natively represented.