folding_schemes::folding::circuits

Module cyclefold

Re-exports

pub use crate::folding::nova::CommittedInstance as CycleFoldCommittedInstance;

pub use crate::folding::nova::Witness as CycleFoldWitness;

Structs

CycleFoldAugmentationGadget

CycleFoldAugmentationGadget implements methods for folding multiple CycleFold instances, both natively and in the augmented step circuit.

CycleFoldChallengeGadget

CycleFoldChallengeGadget computes the RO challenge used for the CycleFold instances NIFS, it contains a rust-native and a in-circuit compatible versions.

CycleFoldCircuit

CycleFoldCommittedInstanceVar

CycleFoldCommittedInstanceVar is the CycleFold CommittedInstance represented in folding verifier circuit

CycleFoldNIFS

CycleFoldNIFS is a wrapper on top of Nova’s NIFS, which just replaces the prove and verify methods to use a different ChallengeGadget, but internally reuses the other Nova’s NIFS methods. It is a custom implementation that does not follow the NIFSTrait because it needs to work over different fields than the main NIFS impls (Nova, Mova, Ova). Could be abstracted, but it’s a tradeoff between overcomplexity at the NIFSTrait and the (not much) need of generalization at the CycleFoldNIFS.

CycleFoldWitnessVar

In-circuit representation of the Witness associated to the CommittedInstance, but with non-native representation, since it is used to represent the CycleFold witness. This struct is used in the Decider circuit.

NIFSFullGadget

This is the gadget used in the AugmentedFCircuit to verify the CycleFold instances folding, which checks the correct RLC of u,x,cmE,cmW (hence the name containing ‘Full’, since it checks all the RLC values, not only the native ones). It assumes that ci2.cmE=0, ci2.u=1.

Traits

CycleFoldConfig

CycleFoldConfig controls the behavior of CycleFoldCircuit.