folding_schemes::folding::circuits::cyclefold

Trait CycleFoldConfig

pub trait CycleFoldConfig<C: Curve>: Sized + Default {
    const N_INPUT_POINTS: usize;
    const N_UNIQUE_RANDOMNESSES: usize;
    const RANDOMNESS_BIT_LENGTH: usize;
    const FIELD_CAPACITY: usize = _;
    const IO_LEN: usize = _;

    // Required methods
    fn alloc_points(
        &self,
        cs: ConstraintSystemRef<CF2<C>>,
    ) -> Result<Vec<C::Var>, SynthesisError>;
    fn alloc_randomnesses(
        &self,
        cs: ConstraintSystemRef<CF2<C>>,
    ) -> Result<Vec<Vec<Boolean<CF2<C>>>>, SynthesisError>;

    // Provided methods
    fn mark_point_as_public(point: &C::Var) -> Result<(), SynthesisError> { ... }
    fn mark_randomness_as_public(
        r: &[Boolean<CF2<C>>],
    ) -> Result<(), SynthesisError> { ... }
    fn build_circuit(self) -> CycleFoldCircuit<C, Self> { ... }
}

CycleFoldConfig controls the behavior of CycleFoldCircuit.

Looking ahead, the circuit computes the random linear combination of points, which is essentially done by iteratively computing P = (P + p_i) * r_i, where P is the folded point, p_i is the input point, and r_i is the randomness.

Required Associated Constants

const N_INPUT_POINTS: usize

N_INPUT_POINTS specifies the number of input points that are folded in CycleFoldCircuit via random linear combinations.

const N_UNIQUE_RANDOMNESSES: usize

N_UNIQUE_RANDOMNESSES specifies the number of unique randomnesses allocated in CycleFoldCircuit. Although the linear combination in general consists of multiple randomnesses, some folding schemes (such as Nova and HyperNova) only need a single one. Thus, by setting this value, the circuit can learn how many randomnesses are used and how long the public inputs vector should be.

const RANDOMNESS_BIT_LENGTH: usize

RANDOMNESS_BIT_LENGTH is the maximum bit length of a randomness r_i.

Provided Associated Constants

const FIELD_CAPACITY: usize = _

FIELD_CAPACITY is the maximum number of bits that can be stored in a field element.

By default, FIELD_CAPACITY is set to MODULUS_BIT_SIZE - 1.

Given a randomness r_i with RANDOMNESS_BIT_LENGTH bits, we need RANDOMNESS_BIT_LENGTH / FIELD_CAPACITY field elements to represent it compactly in-circuit.

const IO_LEN: usize = _

Public inputs length for the CycleFoldCircuit, which depends on the above constants defined by the concrete folding scheme. For example:

In Nova, this is |r| + |p_1| + |p_2| + |P|

In HyperNova, this is |r| + |p_i| * n_points + |P|.

In ProtoGalaxy, this is |[..., r_i, ...]| + |p_i| * n_points + |P|.

As explained above, |r| (i.e., the length of a single randomness) is RANDOMNESS_BIT_LENGTH / FIELD_CAPACITY. When there are multiple randomnesses, the length of |[..., r_i, ...]| is RANDOMNESS_BIT_LENGTH * N_UNIQUE_RANDOMNESSES / FIELD_CAPACITY, as the bits of all randomnesses are concatenated before being packed into field elements. The length of a point p_i when treated as public inputs is 2, as we only need the x and y coordinates of the point.

Thus, IO_LEN is RANDOMNESS_BIT_LENGTH * N_UNIQUE_RANDOMNESSES / FIELD_CAPACITY + 2 * (N_INPUT_POINTS + 1).

Required Methods

fn alloc_points( &self, cs: ConstraintSystemRef<CF2<C>>, ) -> Result<Vec<C::Var>, SynthesisError>

alloc_points allocates the points that are going to be folded in the CycleFoldCircuit via random linear combinations.

The implementation must allocate the points as witness variables (i.e. by calling [AllocVar::new_witness]) first, then mark them as public inputs by calling CycleFoldConfig::mark_point_as_public, and finally return the allocated witness variables.

While it is possible to allocate the points as public inputs directly, we do not use this approach because this will create a longer vector of public inputs, which is not ideal for the augmented step circuit on the primary curve.

fn alloc_randomnesses( &self, cs: ConstraintSystemRef<CF2<C>>, ) -> Result<Vec<Vec<Boolean<CF2<C>>>>, SynthesisError>

alloc_randomnesses allocates the randomnesses used as coefficients of the random linear combinations in the CycleFoldCircuit.

The implementation must allocate the randomnesses as witness variables (i.e. by calling [AllocVar::new_witness]) first, then mark them as public inputs by calling CycleFoldConfig::mark_point_as_public, and finally return the allocated witness variables.

See CycleFoldConfig::alloc_points for the reason why they need to be allocated as witness variables first and converted to public later.

In addition, because the circuit computes P = (P + p_i) * r_i for each i from N_INPUT_POINTS - 1 down to 0, the actual linear combination is P = r_0 * p_0 + (r_0 r_1) * p_1 + (r_0 r_1 r_2) * p_2 + .... Thus, to compute P = R_0 p_0 + R_1 p_1 + R_2 p_2 + ..., the implementation should return r_0 = R_0, r_1 = R_1 / R_0, ..., r_i = R_i / R_{i - 1}. A special case is R_i = R^i, where the allocated randomnesses become r_0 = 1, r_1 = r_2 = ... = R.

Provided Methods

fn mark_point_as_public(point: &C::Var) -> Result<(), SynthesisError>

mark_point_as_public marks a point as public.

The final vector of public inputs is shorter than the result of calling [AllocVar::new_input], because we only need the x and y coordinates of the point, but the infinity flag is not necessary.

fn mark_randomness_as_public( r: &[Boolean<CF2<C>>], ) -> Result<(), SynthesisError>

mark_randomness_as_public marks randomness as public.

The final vector of public inputs is shorter than the result of calling [AllocVar::new_input], because we pack the bits of randomness into a compact field elements.

fn build_circuit(self) -> CycleFoldCircuit<C, Self>

build_circuit creates a new CycleFoldCircuit with self as the configuration.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.