folding_schemes::folding::circuits

Module decider

Modules

off_chain

on_chain

Structs

EvalGadget

Gadget that interpolates the polynomial from the given vector and returns its evaluation at the given point. It also offers the rust native implementation compatible with the gadget.

KZGChallengesGadget

Gadget that computes the KZG challenges. It also offers the rust native implementation compatible with the gadget.

Traits

DeciderEnabledNIFS

This is a temporary workaround for step 6 (running NIFS.V for group elements in circuit) in an NIFS-agnostic way, because different folding schemes have different interfaces of folding verification now.