folding_schemes::folding::circuits::decider::on_chain

Struct GenericOnchainDeciderCircuit

pub struct GenericOnchainDeciderCircuit<C1: Curve, C2: Curve, RU: CommittedInstanceOps<C1>, IU: CommittedInstanceOps<C1>, W: WitnessOps<CF1<C1>>, A: ArithRelation<W, RU>, AVar: ArithRelationGadget<W::Var, RU::Var>, D: DeciderEnabledNIFS<C1, RU, IU, W, A>> {

    pub _avar: PhantomData<AVar>,
    pub arith: A,
    pub cf_arith: R1CS<CF1<C2>>,
    pub cf_pedersen_params: Params<C2>,
    pub poseidon_config: PoseidonConfig<CF1<C1>>,
    pub pp_hash: CF1<C1>,
    pub i: CF1<C1>,
    pub z_0: Vec<CF1<C1>>,
    pub z_i: Vec<CF1<C1>>,
    pub U_i: RU,
    pub W_i: W,
    pub u_i: IU,
    pub w_i: W,
    pub U_i1: RU,
    pub W_i1: W,
    pub proof: D::Proof,
    pub randomness: D::Randomness,
    pub cf_U_i: CycleFoldCommittedInstance<C2>,
    pub cf_W_i: CycleFoldWitness<C2>,
    pub kzg_challenges: Vec<CF1<C1>>,
    pub kzg_evaluations: Vec<CF1<C1>>,

}

A generic circuit tailored for the onchain (Ethereum’s EVM) verification of IVC proofs, where we support IVC built upon any folding scheme.

Specifically, GenericDeciderEthCircuit implements the in-circuit version of the IVC verification algorithm, which essentially checks the following:

R_arith(W_i, U_i): The running instance U_i and witness W_i satisfy arith, and the commitments in U_i open to the values in W_i.

R_arith(w_i, u_i): The incoming instance u_i and witness w_i satisfy arith, and the commitments in u_i open to the values in w_i.

R_cf_arith(cf_W_i, cf_U_i): The CycleFold instance cf_U_i and witness cf_W_i satisfy cf_arith, and the commitments in cf_U_i open to the values in cf_W_i.

u_i contains the correct hash of the initial and final states.

To reduce the number of relation checks, the prover, before invoking the circuit, further folds U_i, u_i into U_{i+1}, and W_i, w_i into W_{i+1}. Now, the circuit only needs to perform two relation checks, i.e., R_arith(W_{i+1}, U_{i+1}) and R_cf_arith(cf_W_i, cf_U_i), plus a few constraints for enforcing the correct hash in u_i and the correct folding from U_i, u_i to U_{i+1}.

We further reduce the circuit size by avoiding the non-native commitment checks involved in R_arith(W_{i+1}, U_{i+1}). Now, we now only check the satisfiability of the constraint system arith with the witness W_{i+1} and instance U_{i+1} in the circuit, but the actual commitment checks are done with the help of KZG.

For more details, see [https://privacy-scaling-explorations.github.io/sonobe-docs/design/nova-decider-onchain.html].

Fields
_avar: PhantomData<AVar>arith: A

Constraint system of the Augmented Function circuit

cf_arith: R1CS<CF1<C2>>

R1CS of the CycleFold circuit

cf_pedersen_params: Params<C2>

CycleFold PedersenParams over C2

poseidon_config: PoseidonConfig<CF1<C1>>pp_hash: CF1<C1>

public params hash

i: CF1<C1>z_0: Vec<CF1<C1>>

initial state

z_i: Vec<CF1<C1>>

current i-th state

U_i: RU

Folding scheme instances

W_i: Wu_i: IUw_i: WU_i1: RUW_i1: Wproof: D::Proof

Helper for folding verification

randomness: D::Randomnesscf_U_i: CycleFoldCommittedInstance<C2>

CycleFold running instance

cf_W_i: CycleFoldWitness<C2>kzg_challenges: Vec<CF1<C1>>

KZG challenges

kzg_evaluations: Vec<CF1<C1>>