folding_schemes::folding::circuits::decider

Trait DeciderEnabledNIFS

pub trait DeciderEnabledNIFS<C: Curve, RU: CommittedInstanceOps<C>, IU: CommittedInstanceOps<C>, W: WitnessOps<CF1<C>>, A: ArithRelation<W, RU>> {
    type ProofDummyCfg;
    type Proof: Dummy<Self::ProofDummyCfg>;
    type RandomnessDummyCfg;
    type Randomness: Dummy<Self::RandomnessDummyCfg>;

    // Required methods
    fn fold_field_elements_gadget(
        arith: &A,
        transcript: &mut PoseidonSpongeVar<CF1<C>>,
        U: RU::Var,
        U_vec: Vec<FpVar<CF1<C>>>,
        u: IU::Var,
        proof: Self::Proof,
        randomness: Self::Randomness,
    ) -> Result<RU::Var, SynthesisError>;
    fn fold_group_elements_native(
        U_commitments: &[C],
        u_commitments: &[C],
        proof: Option<Self::Proof>,
        randomness: Self::Randomness,
    ) -> Result<Vec<C>, Error>;
}

This is a temporary workaround for step 6 (running NIFS.V for group elements in circuit) in an NIFS-agnostic way, because different folding schemes have different interfaces of folding verification now.

In the future, we may introduce a better solution that uses a trait for all folding schemes that specifies their native and in-circuit behaviors.

Required Associated Types

type ProofDummyCfg

type Proof: Dummy<Self::ProofDummyCfg>

type RandomnessDummyCfg

type Randomness: Dummy<Self::RandomnessDummyCfg>

Required Methods

fn fold_field_elements_gadget( arith: &A, transcript: &mut PoseidonSpongeVar<CF1<C>>, U: RU::Var, U_vec: Vec<FpVar<CF1<C>>>, u: IU::Var, proof: Self::Proof, randomness: Self::Randomness, ) -> Result<RU::Var, SynthesisError>

Fold the field elements in U and u inside the circuit.

U_vec is U expressed as a vector of FpVars, which can be reused before or after calling this function to save constraints.

fn fold_group_elements_native( U_commitments: &[C], u_commitments: &[C], proof: Option<Self::Proof>, randomness: Self::Randomness, ) -> Result<Vec<C>, Error>

Fold the group elements (i.e., commitments) in U and u outside the circuit.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.