folding_schemes::folding::circuits::nonnative::uint

Struct NonNativeUintVar

pub struct NonNativeUintVar<F: PrimeField>(pub Vec<LimbVar<F>>);

NonNativeUintVar represents a non-native unsigned integer (BigUint) in the circuit. We apply xJsnark’s techniques for efficient operations on NonNativeUintVar. Note that NonNativeUintVar is different from arkworks’ NonNativeFieldVar in that the latter runs the expensive reduce (align + modulo in our terminology) after each arithmetic operation, while the former only reduces the integer when explicitly called.

Tuple Fields
0: Vec<LimbVar<F>>
Implementations

impl<F: PrimeField> NonNativeUintVar<F>

pub const fn bits_per_limb() -> usize

impl<F: PrimeField> NonNativeUintVar<F>

pub fn enforce_lt(&self, other: &Self) -> Result<(), SynthesisError>

Enforce self to be less than other, where self and other should be aligned. Adapted from https://github.com/akosba/jsnark/blob/0955389d0aae986ceb25affc72edf37a59109250/JsnarkCircuitBuilder/src/circuit/auxiliary/LongElement.java#L801-L872

pub fn enforce_equal_unaligned( &self, other: &Self, ) -> Result<(), SynthesisError>

Enforce self to be equal to other, where self and other are not necessarily aligned.

Adapted from https://github.com/akosba/jsnark/blob/0955389d0aae986ceb25affc72edf37a59109250/JsnarkCircuitBuilder/src/circuit/auxiliary/LongElement.java#L562-L798 Similar implementations can also be found in https://github.com/alex-ozdemir/bellman-bignat/blob/0585b9d90154603a244cba0ac80b9aafe1d57470/src/mp/bignat.rs#L566-L661 and https://github.com/arkworks-rs/r1cs-std/blob/4020fbc22625621baa8125ede87abaeac3c1ca26/src/fields/emulated_fp/reduce.rs#L201-L323

impl<F: PrimeField> NonNativeUintVar<F>

pub fn ubound(&self) -> BigUint

pub fn add_no_align(&self, other: &Self) -> Result<Self, SynthesisError>

Compute self + other, without aligning the limbs.

pub fn mul_no_align(&self, other: &Self) -> Result<Self, SynthesisError>

Compute self * other, without aligning the limbs. Implements the O(n) approach described in xJsnark, Section IV.B.1)

pub fn modulo<M: PrimeField>(&self) -> Result<Self, SynthesisError>

Convert Self to an element in M, i.e., compute Self % M::MODULUS.

pub fn enforce_congruent<M: PrimeField>( &self, other: &Self, ) -> Result<(), SynthesisError>

Enforce that self is congruent to other modulo M::MODULUS.