folding_schemes::folding::nova::nifs

Module ova

Structs

CommittedInstance

A CommittedInstance in Ova is represented by W or W'. It is the result of the commitment to a vector that contains the witness w concatenated with t or e + the public inputs x and a relaxation factor u. (Notice that in the Ova document u is denoted as mu, in this implementation we use u so it follows the original Nova notation, so code is easier to follow).

NIFS

Implements the NIFS (Non-Interactive Folding Scheme) trait for Ova.

Witness

A Witness in Ova is represented by w. It also contains a blinder which can or not be used when committing to the witness itself.

Functions

compute_E

Computes the E parameter (error terms) for the given R1CS and the instance’s z and u. This method is used by the verifier to obtain E in order to check the RelaxedR1CS relation.