folding_schemes::folding::nova::nifs

Trait NIFSTrait

pub trait NIFSTrait<C: Curve, CS: CommitmentScheme<C, H>, T: Transcript<C::ScalarField>, const H: bool = false> {
    type CommittedInstance: Debug + Clone + Absorb;
    type Witness: Debug + Clone;
    type ProverAux: Debug + Clone;
    type Proof: Debug + Clone;

    // Required methods
    fn new_witness(
        w: Vec<C::ScalarField>,
        e_len: usize,
        rng: impl RngCore,
    ) -> Self::Witness;
    fn new_instance(
        rng: impl RngCore,
        params: &CS::ProverParams,
        w: &Self::Witness,
        x: Vec<C::ScalarField>,
        aux: Vec<C::ScalarField>,
    ) -> Result<Self::CommittedInstance, Error>;
    fn fold_witness(
        r: C::ScalarField,
        W: &Self::Witness,
        w: &Self::Witness,
        aux: &Self::ProverAux,
    ) -> Result<Self::Witness, Error>;
    fn prove(
        cs_prover_params: &CS::ProverParams,
        r1cs: &R1CS<C::ScalarField>,
        transcript: &mut T,
        W_i: &Self::Witness,
        U_i: &Self::CommittedInstance,
        w_i: &Self::Witness,
        u_i: &Self::CommittedInstance,
    ) -> Result<(Self::Witness, Self::CommittedInstance, Self::Proof, Vec<bool>), Error>;
    fn verify(
        transcript: &mut T,
        U_i: &Self::CommittedInstance,
        u_i: &Self::CommittedInstance,
        proof: &Self::Proof,
    ) -> Result<(Self::CommittedInstance, Vec<bool>), Error>;
}

Defines the NIFS (Non-Interactive Folding Scheme) trait, initially defined in Nova, and it’s variants Ova and Mova. H specifies whether the NIFS will use a blinding factor.

Required Associated Types

type CommittedInstance: Debug + Clone + Absorb

type Witness: Debug + Clone

type ProverAux: Debug + Clone

type Proof: Debug + Clone

Required Methods

fn new_witness( w: Vec<C::ScalarField>, e_len: usize, rng: impl RngCore, ) -> Self::Witness

fn new_instance( rng: impl RngCore, params: &CS::ProverParams, w: &Self::Witness, x: Vec<C::ScalarField>, aux: Vec<C::ScalarField>, ) -> Result<Self::CommittedInstance, Error>

fn fold_witness( r: C::ScalarField, W: &Self::Witness, w: &Self::Witness, aux: &Self::ProverAux, ) -> Result<Self::Witness, Error>

fn prove( cs_prover_params: &CS::ProverParams, r1cs: &R1CS<C::ScalarField>, transcript: &mut T, W_i: &Self::Witness, U_i: &Self::CommittedInstance, w_i: &Self::Witness, u_i: &Self::CommittedInstance, ) -> Result<(Self::Witness, Self::CommittedInstance, Self::Proof, Vec<bool>), Error>

NIFS.P. Returns a tuple containing the folded Witness, the folded CommittedInstance, and the used challenge r as a vector of bits, so that it can be reused in other methods.

fn verify( transcript: &mut T, U_i: &Self::CommittedInstance, u_i: &Self::CommittedInstance, proof: &Self::Proof, ) -> Result<(Self::CommittedInstance, Vec<bool>), Error>

NIFS.V. Returns the folded CommittedInstance and the used challenge r as a vector of bits, so that it can be reused in other methods.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.