folding_schemes::folding::traits

Trait CommittedInstanceOps

pub trait CommittedInstanceOps<C: Curve>: Inputize<CF1<C>> {
    type Var: AllocVar<Self, CF1<C>> + CommittedInstanceVarOps<C>;

    // Required methods
    fn get_commitments(&self) -> Vec<C>;
    fn is_incoming(&self) -> bool;

    // Provided methods
    fn hash<T: Transcript<CF1<C>>>(
        &self,
        sponge: &T,
        i: CF1<C>,
        z_0: &[CF1<C>],
        z_i: &[CF1<C>],
    ) -> CF1<C>
       where Self: Sized + Absorb { ... }
    fn check_incoming(&self) -> Result<(), Error> { ... }
}

Required Associated Types

type Var: AllocVar<Self, CF1<C>> + CommittedInstanceVarOps<C>

The in-circuit representation of the committed instance.

Required Methods

fn get_commitments(&self) -> Vec<C>

Returns the commitments contained in the committed instance.

fn is_incoming(&self) -> bool

Returns true if the committed instance is an incoming instance, and false if it is a running instance.

Provided Methods

fn hash<T: Transcript<CF1<C>>>( &self, sponge: &T, i: CF1<C>, z_0: &[CF1<C>], z_i: &[CF1<C>], ) -> CF1<C>
where Self: Sized + Absorb,

hash implements the committed instance hash compatible with the in-circuit implementation from CommittedInstanceVarOps::hash.

Returns H(i, z_0, z_i, U_i), where i can be i but also i+1, and U_i is the committed instance self.

fn check_incoming(&self) -> Result<(), Error>

Checks if the committed instance is an incoming instance.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.