folding_schemes::folding::traits

Trait CommittedInstanceVarOps

pub trait CommittedInstanceVarOps<C: Curve> {
    type PointVar: AbsorbNonNativeGadget<CF1<C>>;

    // Required methods
    fn get_commitments(&self) -> Vec<Self::PointVar>;
    fn get_public_inputs(&self) -> &[FpVar<CF1<C>>];
    fn enforce_incoming(&self) -> Result<(), SynthesisError>;
    fn enforce_partial_equal(&self, other: &Self) -> Result<(), SynthesisError>;

    // Provided method
    fn hash(
        &self,
        sponge: &PoseidonSpongeVar<CF1<C>>,
        i: &FpVar<CF1<C>>,
        z_0: &[FpVar<CF1<C>>],
        z_i: &[FpVar<CF1<C>>],
    ) -> Result<(FpVar<CF1<C>>, Vec<FpVar<CF1<C>>>), SynthesisError>
       where Self: AbsorbGadget<CF1<C>> { ... }
}

Required Associated Types

type PointVar: AbsorbNonNativeGadget<CF1<C>>

Required Methods

fn get_commitments(&self) -> Vec<Self::PointVar>

Returns the commitments contained in the committed instance.

fn get_public_inputs(&self) -> &[FpVar<CF1<C>>]

Returns the public inputs contained in the committed instance.

fn enforce_incoming(&self) -> Result<(), SynthesisError>

Generates constraints to enforce that the committed instance is an incoming instance.

fn enforce_partial_equal(&self, other: &Self) -> Result<(), SynthesisError>

Generates constraints to enforce that the committed instance self is partially equal to another committed instance other. Here, only field elements are compared, while commitments (points) are not.

Provided Methods

fn hash( &self, sponge: &PoseidonSpongeVar<CF1<C>>, i: &FpVar<CF1<C>>, z_0: &[FpVar<CF1<C>>], z_i: &[FpVar<CF1<C>>], ) -> Result<(FpVar<CF1<C>>, Vec<FpVar<CF1<C>>>), SynthesisError>
where Self: AbsorbGadget<CF1<C>>,

hash implements the in-circuit committed instance hash compatible with the native implementation from CommittedInstanceOps::hash. Returns H(i, z_0, z_i, U_i), where i can be i but also i+1, and U_i is the committed instance self.

Additionally it returns the in-circuit representation of the committed instance self as a vector of field elements, so they can be reused in other gadgets avoiding recalculating (reconstraining) them.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.