folding_schemes::folding::traits

Trait Inputize

pub trait Inputize<F> {
    // Required method
    fn inputize(&self) -> Vec<F>;
}

Converts a value self into a vector of field elements, ordered in the same way as how a variable of type Var would be represented natively in the circuit.

This is useful for the verifier to compute the public inputs.

Required Methods

fn inputize(&self) -> Vec<F>