folding_schemes::folding::traits

Trait InputizeNonNative

pub trait InputizeNonNative<F> {
    // Required method
    fn inputize_nonnative(&self) -> Vec<F>;
}

Converts a value self into a vector of field elements, ordered in the same way as how a variable of type Var would be represented non-natively in the circuit.

This is useful for the verifier to compute the public inputs.

Note that we require this trait because we need to distinguish between some data types that are represented both natively and non-natively in-circuit (e.g., field elements can have type FpVar and NonNativeUintVar).

Required Methods

fn inputize_nonnative(&self) -> Vec<F>