folding_schemes::frontend

Trait FCircuit

pub trait FCircuit<F: PrimeField>: Clone + Debug {
    type Params: Debug;
    type ExternalInputs: Clone + Default + Debug;
    type ExternalInputsVar: Clone + Debug + AllocVar<Self::ExternalInputs, F>;

    // Required methods
    fn new(params: Self::Params) -> Result<Self, Error>;
    fn state_len(&self) -> usize;
    fn generate_step_constraints(
        &self,
        cs: ConstraintSystemRef<F>,
        i: usize,
        z_i: Vec<FpVar<F>>,
        external_inputs: Self::ExternalInputsVar,
    ) -> Result<Vec<FpVar<F>>, SynthesisError>;
}

FCircuit defines the trait of the circuit of the F function, which is the one being folded (ie. inside the agmented F’ function). The parameter z_i denotes the current state, and z_{i+1} denotes the next state after applying the step. Note that the external inputs for the specific circuit are defined at the implementation of both FCircuit::ExternalInputs and FCircuit::ExternalInputsVar, where the Default trait implementation for the ExternalInputs returns the initialized data structure (ie. if the type contains a vector, it is initialized at the expected length).

Required Associated Types

type Params: Debug

type ExternalInputs: Clone + Default + Debug

type ExternalInputsVar: Clone + Debug + AllocVar<Self::ExternalInputs, F>

Required Methods

fn new(params: Self::Params) -> Result<Self, Error>

returns a new FCircuit instance

fn state_len(&self) -> usize

returns the number of elements in the state of the FCircuit, which corresponds to the FCircuit inputs.

fn generate_step_constraints( &self, cs: ConstraintSystemRef<F>, i: usize, z_i: Vec<FpVar<F>>, external_inputs: Self::ExternalInputsVar, ) -> Result<Vec<FpVar<F>>, SynthesisError>

generates the constraints for the step of F for the given z_i

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.