folding_schemes::frontend

Module utils

Structs

CustomFCircuit

CustomFCircuit is a circuit that has the number of constraints specified in the n_constraints parameter. Note that the generated circuit will have very sparse matrices.

DummyCircuit

DummyCircuit is a circuit that has dummy state whose length is specified in the state_len parameter, without any constraints.

WrapperCircuit

WrapperCircuit is a circuit that wraps any circuit that implements the FCircuit trait. This is used to test the FCircuit.generate_step_constraints method. This is a similar wrapping than the one done in the AugmentedFCircuit, but without adding all the extra constraints of the AugmentedF circuit logic, in order to run lighter tests when we’re not interested in the AugmentedF logic but in the wrapping of the circuits.