folding_schemes

Trait Curve

pub trait Curve:
    CurveGroup<ScalarField: Field, BaseField: Field>
    + AbsorbNonNative
    + Inputize<Self::BaseField>
    + InputizeNonNative<Self::ScalarField> {
    type Var: CurveVar<Self, Self::BaseField>;
}

Curve trait is a wrapper around CurveGroup that also includes the necessary bounds for the curve to be used conveniently in folding schemes.

Required Associated Types

type Var: CurveVar<Self, Self::BaseField>

The in-circuit variable type for this curve.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.