folding_schemes

Trait Field

pub trait Field:
    PrimeField<BasePrimeField = Self>
    + Absorb
    + AbsorbNonNative
    + Inputize<Self> {
    type Var: FieldVar<Self, Self>;
}

Field trait is a wrapper around PrimeField that also includes the necessary bounds for the field to be used conveniently in folding schemes.

Required Associated Types

type Var: FieldVar<Self, Self>

The in-circuit variable type for this field.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.