folding_schemes

Trait FoldingScheme

pub trait FoldingScheme<C1: Curve<BaseField = C2::ScalarField, ScalarField = C2::BaseField>, C2: Curve, FC: FCircuit<C1::ScalarField>>: Clone + Debug {
    type PreprocessorParam: Debug + Clone;
    type ProverParam: Debug + Clone + CanonicalSerialize;
    type VerifierParam: Debug + Clone + CanonicalSerialize;
    type RunningInstance: Debug;
    type IncomingInstance: Debug;
    type MultiCommittedInstanceWithWitness: Debug;
    type CFInstance: Debug;
    type IVCProof: PartialEq + Eq + Clone + Debug + CanonicalSerialize + CanonicalDeserialize;

    // Required methods
    fn pp_deserialize_with_mode<R: Read>(
        reader: R,
        compress: Compress,
        validate: Validate,
        fc_params: FC::Params,
    ) -> Result<Self::ProverParam, Error>;
    fn vp_deserialize_with_mode<R: Read>(
        reader: R,
        compress: Compress,
        validate: Validate,
        fc_params: FC::Params,
    ) -> Result<Self::VerifierParam, Error>;
    fn preprocess(
        rng: impl RngCore,
        prep_param: &Self::PreprocessorParam,
    ) -> Result<(Self::ProverParam, Self::VerifierParam), Error>;
    fn init(
        params: &(Self::ProverParam, Self::VerifierParam),
        step_circuit: FC,
        z_0: Vec<C1::ScalarField>,
    ) -> Result<Self, Error>;
    fn prove_step(
        &mut self,
        rng: impl RngCore,
        external_inputs: FC::ExternalInputs,
        other_instances: Option<Self::MultiCommittedInstanceWithWitness>,
    ) -> Result<(), Error>;
    fn state(&self) -> Vec<C1::ScalarField>;
    fn ivc_proof(&self) -> Self::IVCProof;
    fn from_ivc_proof(
        ivc_proof: Self::IVCProof,
        fcircuit_params: FC::Params,
        params: (Self::ProverParam, Self::VerifierParam),
    ) -> Result<Self, Error>;
    fn verify(
        vp: Self::VerifierParam,
        ivc_proof: Self::IVCProof,
    ) -> Result<(), Error>;
}

FoldingScheme defines trait that is implemented by the diverse folding schemes. It is defined over a cycle of curves (C1, C2), where:

C1 is the main curve, which ScalarField we use as our F for all the field operations

C2 is the auxiliary curve, which we use for the commitments, whose BaseField (for point coordinates) are in the C1::ScalarField.

In other words, C1.Fq == C2.Fr, and C1.Fr == C2.Fq.

Required Associated Types

type PreprocessorParam: Debug + Clone

type ProverParam: Debug + Clone + CanonicalSerialize

type VerifierParam: Debug + Clone + CanonicalSerialize

type RunningInstance: Debug

type IncomingInstance: Debug

type MultiCommittedInstanceWithWitness: Debug

type CFInstance: Debug

type IVCProof: PartialEq + Eq + Clone + Debug + CanonicalSerialize + CanonicalDeserialize

Required Methods

fn pp_deserialize_with_mode<R: Read>( reader: R, compress: Compress, validate: Validate, fc_params: FC::Params, ) -> Result<Self::ProverParam, Error>

deserialize Self::ProverParam and recover the not serialized data that is recomputed on the fly to save serialized bytes. Internally it generates the r1cs/ccs & cf_r1cs needed for the VerifierParams. In this way we avoid needing to serialize them, saving significant space in the VerifierParams serialized size.

fn vp_deserialize_with_mode<R: Read>( reader: R, compress: Compress, validate: Validate, fc_params: FC::Params, ) -> Result<Self::VerifierParam, Error>

deserialize Self::VerifierParam and recover the not serialized data that is recomputed on the fly to save serialized bytes. Internally it generates the r1cs/ccs & cf_r1cs needed for the VerifierParams. In this way we avoid needing to serialize them, saving significant space in the VerifierParams serialized size.

fn preprocess( rng: impl RngCore, prep_param: &Self::PreprocessorParam, ) -> Result<(Self::ProverParam, Self::VerifierParam), Error>

fn init( params: &(Self::ProverParam, Self::VerifierParam), step_circuit: FC, z_0: Vec<C1::ScalarField>, ) -> Result<Self, Error>

fn prove_step( &mut self, rng: impl RngCore, external_inputs: FC::ExternalInputs, other_instances: Option<Self::MultiCommittedInstanceWithWitness>, ) -> Result<(), Error>

fn state(&self) -> Vec<C1::ScalarField>

returns the state at the current step

fn ivc_proof(&self) -> Self::IVCProof

returns the last IVC state proof, which can be verified in the verify method

fn from_ivc_proof( ivc_proof: Self::IVCProof, fcircuit_params: FC::Params, params: (Self::ProverParam, Self::VerifierParam), ) -> Result<Self, Error>

constructs the FoldingScheme instance from the given IVCProof, ProverParams, VerifierParams and PoseidonConfig. This method is useful for when the IVCProof is sent between different parties, so that they can continue iterating the IVC from the received IVCProof.

fn verify( vp: Self::VerifierParam, ivc_proof: Self::IVCProof, ) -> Result<(), Error>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.