folding_schemes

Trait MultiFolding

pub trait MultiFolding<C1: Curve<BaseField = C2::ScalarField, ScalarField = C2::BaseField>, C2: Curve, FC: FCircuit<C1::ScalarField>>: Clone + Debug {
    type RunningInstance: Debug;
    type IncomingInstance: Debug;
    type MultiInstance: Debug;

    // Required methods
    fn new_running_instance(
        &self,
        rng: impl RngCore,
        state: Vec<C1::ScalarField>,
        external_inputs: FC::ExternalInputs,
    ) -> Result<Self::RunningInstance, Error>;
    fn new_incoming_instance(
        &self,
        rng: impl RngCore,
        state: Vec<C1::ScalarField>,
        external_inputs: FC::ExternalInputs,
    ) -> Result<Self::IncomingInstance, Error>;
}

Trait with auxiliary methods for multi-folding schemes (ie. HyperNova, ProtoGalaxy, etc), allowing to create new instances for the multifold.

Required Associated Types

type RunningInstance: Debug

type IncomingInstance: Debug

type MultiInstance: Debug

Required Methods

fn new_running_instance( &self, rng: impl RngCore, state: Vec<C1::ScalarField>, external_inputs: FC::ExternalInputs, ) -> Result<Self::RunningInstance, Error>

Creates a new RunningInstance for the given state, to be folded in the multi-folding step.

fn new_incoming_instance( &self, rng: impl RngCore, state: Vec<C1::ScalarField>, external_inputs: FC::ExternalInputs, ) -> Result<Self::IncomingInstance, Error>

Creates a new IncomingInstance for the given state, to be folded in the multi-folding step.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.