folding_schemes::transcript

Trait AbsorbNonNative

pub trait AbsorbNonNative {
    // Required method
    fn to_native_sponge_field_elements<F: PrimeField>(&self, dest: &mut Vec<F>);

    // Provided method
    fn to_native_sponge_field_elements_as_vec<F: PrimeField>(&self) -> Vec<F> { ... }
}

An interface for objects that can be absorbed by a Transcript.

Matches Absorb in ark-crypto-primitives.

Required Methods

fn to_native_sponge_field_elements<F: PrimeField>(&self, dest: &mut Vec<F>)

Converts the object into field elements that can be absorbed by a Transcript. Append the list to dest

Provided Methods

fn to_native_sponge_field_elements_as_vec<F: PrimeField>(&self) -> Vec<F>

Converts the object into field elements that can be absorbed by a Transcript. Return the list as Vec

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.