folding_schemes::transcript

Trait AbsorbNonNativeGadget

pub trait AbsorbNonNativeGadget<F: PrimeField> {
    // Required method
    fn to_native_sponge_field_elements(
        &self,
    ) -> Result<Vec<FpVar<F>>, SynthesisError>;
}

An interface for objects that can be absorbed by a TranscriptVar whose constraint field is F.

Matches AbsorbGadget in ark-crypto-primitives.

Required Methods

fn to_native_sponge_field_elements( &self, ) -> Result<Vec<FpVar<F>>, SynthesisError>

Converts the object into field elements that can be absorbed by a TranscriptVar.