folding_schemes::transcript

Trait Transcript

pub trait Transcript<F: PrimeField>: CryptographicSponge {
    // Required methods
    fn new_with_pp_hash(config: &Self::Config, pp_hash: F) -> Self;
    fn absorb_point<C: CurveGroup<BaseField = F>>(&mut self, v: &C);
    fn absorb_nonnative<V: AbsorbNonNative>(&mut self, v: &V);
    fn get_challenge(&mut self) -> F;
    fn get_challenge_nbits(&mut self, nbits: usize) -> Vec<bool>;
    fn get_challenges(&mut self, n: usize) -> Vec<F>;
}

Required Methods

fn new_with_pp_hash(config: &Self::Config, pp_hash: F) -> Self

new_with_pp_hash creates a new transcript / sponge with the given hash of the public parameters.

fn absorb_point<C: CurveGroup<BaseField = F>>(&mut self, v: &C)

absorb_point is for absorbing points whose BaseField is the field of the sponge, i.e., the type C of these points should satisfy C::BaseField = F.

If the sponge field F is C::ScalarField, call absorb_nonnative instead.

fn absorb_nonnative<V: AbsorbNonNative>(&mut self, v: &V)

absorb_nonnative is for structs that contain non-native (field or group) elements, including:

A field element of type T: PrimeField that will be absorbed into a sponge that operates in another field F != T.

A group element of type C: CurveGroup that will be absorbed into a sponge that operates in another field F != C::BaseField, e.g., F = C::ScalarField.

A CommittedInstance on the secondary curve (used for CycleFold) that will be absorbed into a sponge that operates in the (scalar field of the) primary curve.

Note that although a CommittedInstance for AugmentedFCircuit on the primary curve also contains non-native elements, we still regard it as native, because the sponge is on the same curve.

fn get_challenge(&mut self) -> F

fn get_challenge_nbits(&mut self, nbits: usize) -> Vec<bool>

get_challenge_nbits returns a field element of size nbits

fn get_challenges(&mut self, n: usize) -> Vec<F>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.