folding_schemes::transcript

Trait TranscriptVar

pub trait TranscriptVar<F: PrimeField, S: CryptographicSponge>: CryptographicSpongeVar<F, S> {
    // Required methods
    fn new_with_pp_hash(
        config: &Self::Parameters,
        pp_hash: &FpVar<F>,
    ) -> Result<Self, SynthesisError>;
    fn absorb_point<C: CurveGroup<BaseField = F>, GC: CurveVar<C, F>>(
        &mut self,
        v: &GC,
    ) -> Result<(), SynthesisError>;
    fn absorb_nonnative<V: AbsorbNonNativeGadget<F>>(
        &mut self,
        v: &V,
    ) -> Result<(), SynthesisError>;
    fn get_challenge(&mut self) -> Result<FpVar<F>, SynthesisError>;
    fn get_challenge_nbits(
        &mut self,
        nbits: usize,
    ) -> Result<Vec<Boolean<F>>, SynthesisError>;
    fn get_challenges(
        &mut self,
        n: usize,
    ) -> Result<Vec<FpVar<F>>, SynthesisError>;
}

Required Methods

fn new_with_pp_hash( config: &Self::Parameters, pp_hash: &FpVar<F>, ) -> Result<Self, SynthesisError>

new_with_pp_hash creates a new transcript / sponge with the given hash of the public parameters.

fn absorb_point<C: CurveGroup<BaseField = F>, GC: CurveVar<C, F>>( &mut self, v: &GC, ) -> Result<(), SynthesisError>

absorb_point is for absorbing points whose BaseField is the field of the sponge, i.e., the type C of these points should satisfy C::BaseField = F.

If the sponge field F is C::ScalarField, call absorb_nonnative instead.

fn absorb_nonnative<V: AbsorbNonNativeGadget<F>>( &mut self, v: &V, ) -> Result<(), SynthesisError>

absorb_nonnative is for structs that contain non-native (field or group) elements, including:

A field element of type T: PrimeField that will be absorbed into a sponge that operates in another field F != T.

A group element of type C: CurveGroup that will be absorbed into a sponge that operates in another field F != C::BaseField, e.g., F = C::ScalarField.

A CommittedInstance on the secondary curve (used for CycleFold) that will be absorbed into a sponge that operates in the (scalar field of the) primary curve.

Note that although a CommittedInstance for AugmentedFCircuit on the primary curve also contains non-native elements, we still regard it as native, because the sponge is on the same curve.

fn get_challenge(&mut self) -> Result<FpVar<F>, SynthesisError>

fn get_challenge_nbits( &mut self, nbits: usize, ) -> Result<Vec<Boolean<F>>, SynthesisError>

returns the bit representation of the challenge, we use its output in-circuit for the GC.scalar_mul_le method.

fn get_challenges(&mut self, n: usize) -> Result<Vec<FpVar<F>>, SynthesisError>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.