folding_schemes::utils::espresso

Module sum_check

This module implements the sum check protocol.

Modules

structs

This module defines structs that are shared by all sub protocols.

verifier

Verifier subroutines for a SumCheck protocol.

Structs

IOPSumCheck

SumCheckSubClaim

A SumCheckSubClaim is a claim generated by the verifier at the end of verification when it is convinced.

Traits

SumCheck

A generic sum-check trait over a curve group

SumCheckProver

Trait for sum check protocol prover side APIs.

SumCheckVerifier

Trait for sum check protocol verifier side APIs.