folding_schemes::utils::espresso::sum_check

Trait SumCheck

pub trait SumCheck<F: PrimeField> {
    type VirtualPolynomial;
    type VPAuxInfo;
    type MultilinearExtension;
    type SumCheckProof: Clone + Debug + Default + PartialEq;
    type SumCheckSubClaim: Clone + Debug + Default + PartialEq;

    // Required methods
    fn extract_sum(proof: &Self::SumCheckProof) -> F;
    fn prove(
        poly: &Self::VirtualPolynomial,
        transcript: &mut impl Transcript<F>,
    ) -> Result<Self::SumCheckProof, Error>;
    fn verify(
        sum: F,
        proof: &Self::SumCheckProof,
        aux_info: &Self::VPAuxInfo,
        transcript: &mut impl Transcript<F>,
    ) -> Result<Self::SumCheckSubClaim, Error>;
}

A generic sum-check trait over a curve group

Required Associated Types

type VirtualPolynomial

type VPAuxInfo

type MultilinearExtension

type SumCheckProof: Clone + Debug + Default + PartialEq

type SumCheckSubClaim: Clone + Debug + Default + PartialEq

Required Methods

fn extract_sum(proof: &Self::SumCheckProof) -> F

Extract sum from the proof

fn prove( poly: &Self::VirtualPolynomial, transcript: &mut impl Transcript<F>, ) -> Result<Self::SumCheckProof, Error>

Generate proof of the sum of polynomial over {0,1}^num_vars

The polynomial is represented in the form of a VirtualPolynomial.

fn verify( sum: F, proof: &Self::SumCheckProof, aux_info: &Self::VPAuxInfo, transcript: &mut impl Transcript<F>, ) -> Result<Self::SumCheckSubClaim, Error>

Verify the claimed sum using the proof

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.