folding_schemes::utils::espresso::sum_check

Trait SumCheckProver

pub trait SumCheckProver<F: PrimeField>: Sized {
    type VirtualPolynomial;
    type ProverMessage;

    // Required methods
    fn prover_init(polynomial: &Self::VirtualPolynomial) -> Result<Self, Error>;
    fn prove_round_and_update_state(
        &mut self,
        challenge: &Option<F>,
    ) -> Result<Self::ProverMessage, Error>;
}

Trait for sum check protocol prover side APIs.

Required Associated Types

type VirtualPolynomial

type ProverMessage

Required Methods

fn prover_init(polynomial: &Self::VirtualPolynomial) -> Result<Self, Error>

Initialize the prover state to argue for the sum of the input polynomial over {0,1}^num_vars.

fn prove_round_and_update_state( &mut self, challenge: &Option<F>, ) -> Result<Self::ProverMessage, Error>

Receive message from verifier, generate prover message, and proceed to next round.

Main algorithm used is from section 3.2 of XZZPS19.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.