folding_schemes::utils::espresso::sum_check

Trait SumCheckVerifier

pub trait SumCheckVerifier<F: PrimeField> {
    type VPAuxInfo;
    type ProverMessage;
    type Challenge;
    type SumCheckSubClaim;

    // Required methods
    fn verifier_init(index_info: &Self::VPAuxInfo) -> Self;
    fn verify_round_and_update_state(
        &mut self,
        prover_msg: &Self::ProverMessage,
        transcript: &mut impl Transcript<F>,
    ) -> Result<Self::Challenge, Error>;
    fn check_and_generate_subclaim(
        &self,
        asserted_sum: &F,
    ) -> Result<Self::SumCheckSubClaim, Error>;
}

Trait for sum check protocol verifier side APIs.

Required Associated Types

type VPAuxInfo

type ProverMessage

type Challenge

type SumCheckSubClaim

Required Methods

fn verifier_init(index_info: &Self::VPAuxInfo) -> Self

Initialize the verifier’s state.

fn verify_round_and_update_state( &mut self, prover_msg: &Self::ProverMessage, transcript: &mut impl Transcript<F>, ) -> Result<Self::Challenge, Error>

Run verifier for the current round, given a prover message.

Note that verify_round_and_update_state only samples and stores challenges; and update the verifier’s state accordingly. The actual verifications are deferred (in batch) to check_and_generate_subclaim at the last step.

fn check_and_generate_subclaim( &self, asserted_sum: &F, ) -> Result<Self::SumCheckSubClaim, Error>

This function verifies the deferred checks in the interactive version of the protocol; and generate the subclaim. Returns an error if the proof failed to verify.

If the asserted sum is correct, then the multilinear polynomial evaluated at subclaim.point will be subclaim.expected_evaluation. Otherwise, it is highly unlikely that those two will be equal. Larger field size guarantees smaller soundness error.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.