folding_schemes::utils

Module mle

Functions

dense_vec_to_dense_mle

Takes the n_vars and a dense vector and returns its dense MLE.

dense_vec_to_mle

Takes the n_vars and a dense vector and returns its sparse MLE.

matrix_to_dense_mle

Returns the dense multilinear extension from the given matrix, without modifying the original matrix.

matrix_to_mle

Returns the sparse multilinear extension from the given matrix, without modifying the original matrix.

pad_matrix

Pad matrix so that its columns and rows are powers of two

vec_to_dense_mle

Takes the n_vars and a dense vector and returns its dense MLE.

vec_to_mle

Takes the n_vars and a sparse vector and returns its sparse MLE.